Lenny supports api keys that when in use requires requests to use this key to authorize the requests
The api key must be configured in the internal service properties

+++ src/org/springfield/lenny/restlet/LennyResource.java
@@ -38,8 +38,10 @@ import org.dom4j.Node;
 import org.restlet.Context;
 import org.restlet.Request;
 import org.restlet.Response;
+import org.restlet.data.Form;
 import org.restlet.data.MediaType;
 import org.restlet.data.Status;
+import org.restlet.engine.header.Header;
 import org.restlet.representation.Representation;
 import org.restlet.representation.Variant;
 import org.restlet.resource.Delete;
@@ -47,6 +49,8 @@ import org.restlet.resource.Get;
 import org.restlet.resource.Post;
 import org.restlet.resource.Put;
 import org.restlet.resource.ServerResource;
+import org.restlet.util.Series;
+import org.springfield.lenny.LennyServer;
 import org.springfield.lenny.conditionalaccess.AccessList;
 import org.springfield.lenny.conditionalaccess.AccessListEntry;
 import org.springfield.lenny.restlet.util.Pair;
@@ -70,6 +74,14 @@ public class LennyResource extends ServerResource {
 	
 	private static AccessList accesslist;
 	private List<InetAddress> whitelist = new ArrayList<InetAddress>();
+	
+	/** api key */
+	private static String apiKey;
+	
+	// load static variables for the configuration
+	static {
+		apiKey = LennyServer.instance().getConfiguration().getProperty("apiKey");
+	}
 	
 	public LennyResource() {
 		//constructor
@@ -104,19 +116,24 @@ public class LennyResource extends ServerResource {
         String responseBody;
         MediaType mediatype = MediaType.TEXT_HTML;
         
-        if (uri.equals(TICKET_URI)) {
-			Pair<Status, String> response = showList();
-			status = response.status;
-			responseBody = response.response;			
-		} else if (uri.startsWith(TICKET_URI)) {
-			Pair<Status, String> response = getTicket(uri);
-			status = response.status;
-			responseBody = response.response;
-			mediatype = MediaType.TEXT_XML;
-		} else {
-			status = Status.CLIENT_ERROR_NOT_FOUND; 
-			responseBody = "404 - Not found";
-		}
+        if(!isAuthorized()) {
+        	status = Status.CLIENT_ERROR_FORBIDDEN;
+        	responseBody = "403 - Not authorized";	
+        } else {        
+	        if (uri.equals(TICKET_URI)) {
+				Pair<Status, String> response = showList();
+				status = response.status;
+				responseBody = response.response;			
+			} else if (uri.startsWith(TICKET_URI)) {
+				Pair<Status, String> response = getTicket(uri);
+				status = response.status;
+				responseBody = response.response;
+				mediatype = MediaType.TEXT_XML;
+			} else {
+				status = Status.CLIENT_ERROR_NOT_FOUND; 
+				responseBody = "404 - Not found";
+			}
+        }
 
         getResponse().setStatus(status);
 		getResponse().setEntity(responseBody, mediatype);
@@ -134,28 +151,33 @@ public class LennyResource extends ServerResource {
         String responseBody;
         MediaType mediatype = MediaType.TEXT_HTML;
         
-        String xml = "";
-        
-        try {
-			if (representation == null) {
-				getResponse().setStatus(Status.CLIENT_ERROR_BAD_REQUEST);
-				getResponse().setEntity("<status>Error: the request data could not be read</status>",
-						MediaType.TEXT_XML);
-			} else {
-				xml = representation.getText();
+        if(!isAuthorized()) {
+        	status = Status.CLIENT_ERROR_FORBIDDEN;
+        	responseBody = "403 - Not authorized";	
+        } else {        
+	        String xml = "";
+	        
+	        try {
+				if (representation == null) {
+					getResponse().setStatus(Status.CLIENT_ERROR_BAD_REQUEST);
+					getResponse().setEntity("<status>Error: the request data could not be read</status>",
+							MediaType.TEXT_XML);
+				} else {
+					xml = representation.getText();
+				}
+			} catch (IOException e2) {
+				e2.printStackTrace();
+				return;
 			}
-		} catch (IOException e2) {
-			e2.printStackTrace();
-			return;
-		}
-        
-        if (uri.startsWith(TICKET_ACCESS_URI)) {
-        	Pair<Status, String> response = getTicketHasAccess(uri, xml);
-        	status = response.status;
-        	responseBody = response.response;
-        } else {
-        	status = Status.CLIENT_ERROR_NOT_FOUND; 
-			responseBody = "404 - Not found";
+	        
+	        if (uri.startsWith(TICKET_ACCESS_URI)) {
+	        	Pair<Status, String> response = getTicketHasAccess(uri, xml);
+	        	status = response.status;
+	        	responseBody = response.response;
+	        } else {
+	        	status = Status.CLIENT_ERROR_NOT_FOUND; 
+				responseBody = "404 - Not found";
+	        }
         }
         
         getResponse().setStatus(status);
@@ -174,29 +196,34 @@ public class LennyResource extends ServerResource {
         String responseBody;
         MediaType mediatype = MediaType.TEXT_HTML;
         
-        String xml = "";
-        
-        try {
-			if (representation == null) {
-				getResponse().setStatus(Status.CLIENT_ERROR_BAD_REQUEST);
-				getResponse().setEntity("<status>Error: the request data could not be read</status>",
-						MediaType.TEXT_XML);
-			} else {
-				xml = representation.getText();
+        if(!isAuthorized()) {
+        	status = Status.CLIENT_ERROR_FORBIDDEN;
+        	responseBody = "403 - Not authorized";	
+        } else {          
+	        String xml = "";
+	        
+	        try {
+				if (representation == null) {
+					getResponse().setStatus(Status.CLIENT_ERROR_BAD_REQUEST);
+					getResponse().setEntity("<status>Error: the request data could not be read</status>",
+							MediaType.TEXT_XML);
+				} else {
+					xml = representation.getText();
+				}
+			} catch (IOException e2) {
+				e2.printStackTrace();
+				return;
 			}
-		} catch (IOException e2) {
-			e2.printStackTrace();
-			return;
-		}
-        
-        if (uri.equals(TICKET_URI)) {
-        	Pair<Status, String> response = addTicket(xml);
-        	status = response.status;
-        	responseBody = response.response;
-        	mediatype = MediaType.TEXT_XML;
-        } else {
-        	status = Status.CLIENT_ERROR_NOT_FOUND; 
-			responseBody = "404 - Not found";
+	        
+	        if (uri.equals(TICKET_URI)) {
+	        	Pair<Status, String> response = addTicket(xml);
+	        	status = response.status;
+	        	responseBody = response.response;
+	        	mediatype = MediaType.TEXT_XML;
+	        } else {
+	        	status = Status.CLIENT_ERROR_NOT_FOUND; 
+				responseBody = "404 - Not found";
+	        }
         }
 		
         getResponse().setStatus(status);
@@ -215,14 +242,19 @@ public class LennyResource extends ServerResource {
         String responseBody;
         MediaType mediatype = MediaType.TEXT_HTML;
         
-        if (uri.startsWith(TICKET_URI)) {
-        	Pair<Status, String> response = deleteTicket(uri);
-        	status = response.status;
-        	responseBody = response.response;
-        	mediatype = MediaType.TEXT_XML;
-        } else {
-        	status = Status.CLIENT_ERROR_NOT_FOUND; 
-			responseBody = "404 - Not found";
+        if(!isAuthorized()) {
+        	status = Status.CLIENT_ERROR_FORBIDDEN;
+        	responseBody = "403 - Not authorized";	
+        } else {          
+	        if (uri.startsWith(TICKET_URI)) {
+	        	Pair<Status, String> response = deleteTicket(uri);
+	        	status = response.status;
+	        	responseBody = response.response;
+	        	mediatype = MediaType.TEXT_XML;
+	        } else {
+	        	status = Status.CLIENT_ERROR_NOT_FOUND; 
+				responseBody = "404 - Not found";
+	        }
         }
         
         getResponse().setStatus(status);
@@ -488,18 +520,28 @@ public class LennyResource extends ServerResource {
 		return fsxml;
 	}
 	
-	/*private Document getNode(String uri) {
-		Document response = null;
-		
-		String responseStr = HttpHelper.sendRequest("GET", uri, null, null);
-		
-		try {	
-			response = DocumentHelper.parseText(responseStr);
-		} catch (DocumentException e) {
-			e.printStackTrace();
-			logger.error("could not create document from node "+uri+" - "+e.getMessage());
+	/* 
+	 * Check if request is authorized when an api key is set
+	 */
+	private boolean isAuthorized() {
+		//no api key configured
+		if (apiKey.equals("")) {
+			return true;
 		}
-		return response;
-	}*/
-	
+		
+		//check for x-api-key header
+		Series<Header> headers = (Series<Header>)getRequestAttributes().get("org.restlet.http.headers");
+		String requestHeaderApiKey = headers.getFirstValue("x-api-key", true);
+		
+		//check for api_key parameter		
+		String requestGetParamApiKey = getQueryValue("api_key");
+		
+		if (requestHeaderApiKey != null && requestHeaderApiKey.equals(apiKey)) {
+			return true;
+		} else if (requestGetParamApiKey != null && requestGetParamApiKey.equals(apiKey)) {
+			return true;
+		}
+		
+		return false;
+	}	
 }
